backend/requirements.txt
fastapi
uvicorn
supabase
apscheduler
feedparser
praw
requests
python-dotenv
pydantic
